LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing-01.html

<!DOCTYPE html>
<html>
<head>
<script src='resources/source-list-parsing-base.js'></script>
<script>
var tests = [
    ['yes', 'script-src 127.0.0.1:8000'],
    ['yes', 'script-src http://127.0.0.1:8000'],
    ['no',  'script-src https://127.0.0.1:8000'],
    ['no',  'script-src https://127.0.0.1:*'],
    ['no',  'script-src https://127.0.0.1:*  ']
];
</script>
</head>
<body onload="test()">
  <p>
    None of these scripts should execute even though there are parse errors in the policy.
  </p>
